Refactor sorting tests and README update
