--- src/main/java/com/algdat/tests/sortingtests/SelectionSortTest.java
@@ -1,12 +1,0 @@
-package com.algdat.tests.sortingtests;
-
-import com.algdat.algorithms.sorting.SelectionSort;
-
-class SelectionSortTest {
-    public static void main(String[] args) {
-        SortingTest test = new SortingTest(100, 10, 0, 100);
-        SelectionSort selectionSort = new SelectionSort();
-
-        test.testAlgorithm(selectionSort);
-    }
-}
